requirements.txt
streamlit
numpy
pandas
Pillow
matplotlib
seaborn
plotly
tensorflow
scikit-learn
joblib
requests